+++ src/app/(main)/admin/equipment/page.tsx
@@ -1,11 +1,11 @@
-import Materials from "@/components/Materials";
+import Materials from "@/components/MaterialContainer";
 
 const Equipments = () => {
   return (
-        <div id="laboratory-materials" className="section">
-            <p className="text-white text-center">ADMIN</p>
-        <Materials user_type="admin" />
-        </div>
+    <div id="laboratory-materials" className="section">
+      <p className="text-white text-center">ADMIN</p>
+      <Materials user_type="admin" />
+    </div>
   );
 };
