+++ src/app/(main)/admin/masterlist/page.tsx
@@ -1,11 +1,11 @@
-import Materials from "@/components/Materials";
+import Materials from "@/components/MaterialContainer";
 
 const MasterList = () => {
   return (
-        <div id="laboratory-materials" className="section">
-            <p className="text-white text-center">ADMIN</p>
-        <Materials />
-        </div>
+    <div id="laboratory-materials" className="section">
+      <p className="text-white text-center">ADMIN</p>
+      <Materials />
+    </div>
   );
 };
