+++ src/app/(main)/equipment/page.tsx
@@ -1,13 +1,15 @@
-'use client';
-import Materials from "@/components/Materials";
+"use client";
+import Materials from "@/components/MaterialContainer";
 import Link from "next/link";
 
 const Home = () => {
   return (
-        <div id="laboratory-materials" className="section bg-red-500">
-        <Materials user_type="borrower" />
-        <Link href="/cart" className="hover:underline text-center">Cart</Link>
-        </div>
+    <div id="laboratory-materials" className="section bg-red-500">
+      <Materials user_type="borrower" />
+      <Link href="/cart" className="hover:underline text-center">
+        Cart
+      </Link>
+    </div>
   );
 };
